# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sriswa Studio e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount?: number;
}

export interface CategoryInput {
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ProductImage {
  id: number;
  productId: number;
  url: string;
  isPrimary: boolean;
  displayOrder: number;
}

export interface ProductVariant {
  id: number;
  productId: number;
  name: string;
  value: string;
  priceModifier: number;
  stockQuantity: number;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName?: string | null;
  isFeatured: boolean;
  isActive: boolean;
  stockQuantity: number;
  lowStockThreshold?: number;
  images: ProductImage[];
  variants: ProductVariant[];
  createdAt?: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
}

export interface ProductInput {
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @nullable */
  categoryId?: number | null;
  isFeatured?: boolean;
  isActive?: boolean;
  stockQuantity?: number;
  lowStockThreshold?: number;
}

export interface ProductUpdate {
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @nullable */
  categoryId?: number | null;
  isFeatured?: boolean;
  isActive?: boolean;
  stockQuantity?: number;
  lowStockThreshold?: number;
}

export interface ProductImageInput {
  url: string;
  isPrimary?: boolean;
  displayOrder?: number;
}

export interface Inventory {
  productId: number;
  stockQuantity: number;
  lowStockThreshold: number;
}

export interface InventoryUpdate {
  stockQuantity: number;
  lowStockThreshold?: number;
}

export interface InventoryItem {
  productId: number;
  productName: string;
  stockQuantity: number;
  lowStockThreshold: number;
  isLowStock: boolean;
}

export interface LowStockItem {
  productId: number;
  productName: string;
  stockQuantity: number;
  lowStockThreshold: number;
}

export interface CartItem {
  id: number;
  sessionId: string;
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  /** Total stock units available for this product */
  stockQuantity: number;
  /** Maximum quantity allowed for this specific cart line, accounting for other variant lines of the same product already in the cart */
  maxQuantity: number;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  variantId?: number | null;
  /** @nullable */
  variantLabel?: string | null;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  sessionId: string;
  productId: number;
  quantity: number;
  /** @nullable */
  variantId?: number | null;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  variantLabel?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  status: OrderStatus;
  customerName: string;
  customerPhone: string;
  customerEmail: string;
  shippingAddress: string;
  city?: string;
  state?: string;
  pincode?: string;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  subtotal: number;
  shippingCost?: number;
  total: number;
  paymentMethod?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
}

export interface OrderInput {
  sessionId: string;
  customerName: string;
  customerPhone: string;
  customerEmail: string;
  shippingAddress: string;
  city: string;
  state: string;
  pincode: string;
  /** @nullable */
  notes?: string | null;
  paymentMethod?: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface OrderStatusCount {
  status: string;
  count: number;
}

export interface AdminDashboard {
  todayOrders: number;
  todayRevenue: number;
  totalOrders: number;
  totalRevenue: number;
  lowStockCount: number;
  recentOrders: Order[];
  ordersByStatus: OrderStatusCount[];
}

export interface StoreSettings {
  id: number;
  storeName: string;
  adminWhatsapp: string;
  newOrderTemplate?: string;
  statusUpdateTemplate?: string;
}

export interface StoreSettingsUpdate {
  storeName?: string;
  adminWhatsapp?: string;
  newOrderTemplate?: string;
  statusUpdateTemplate?: string;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListProductsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
minPrice?: number | null;
/**
 * @nullable
 */
maxPrice?: number | null;
/**
 * @nullable
 */
sortBy?: ListProductsSortBy;
/**
 * @nullable
 */
featured?: boolean | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
offset?: number | null;
};

export type ListProductsSortBy = typeof ListProductsSortBy[keyof typeof ListProductsSortBy] | null;


export const ListProductsSortBy = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  newest: 'newest',
  name: 'name',
} as const;

export type GetFeaturedProductsParams = {
/**
 * @nullable
 */
limit?: number | null;
};

export type GetCartParams = {
sessionId: string;
};

export type ClearCartParams = {
sessionId: string;
};

export type ListOrdersParams = {
/**
 * @nullable
 */
status?: ListOrdersStatus;
/**
 * @nullable
 */
dateFrom?: string | null;
/**
 * @nullable
 */
dateTo?: string | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
offset?: number | null;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus] | null;


export const ListOrdersStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type TrackOrderParams = {
/**
 * @nullable
 */
orderNumber?: string | null;
/**
 * @nullable
 */
email?: string | null;
};