# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sriswa Studio e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart for current session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/clear:
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear all items from cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cleared

  # ── ORDERS ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CartItem:
      type: object
      required: [id, sessionId, productId, productName, price, quantity, stockQuantity, maxQuantity, imageUrl]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
        stockQuantity:
          type: integer
          description: Total stock units available for this product
        maxQuantity:
          type: integer
          description: Maximum quantity allowed for this specific cart line, accounting for other variant lines of the same product already in the cart
        imageUrl:
          type: ["string", "null"]
        variantId:
          type: ["integer", "null"]
        variantLabel:
          type: ["string", "null"]
    Cart:
      type: object
      required: [sessionId, items, total, itemCount]
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
    CartItemInput:
      type: object
      required: [sessionId, productId, quantity]
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
        variantId:
          type: ["integer", "null"]
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    # ── ORDER ─────────────────────────────────────────────────────────────────
...