# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sriswa Studio e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: number;
  sessionId: string;
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  /** Total stock units available for this product */
  stockQuantity: number;
  /** Maximum quantity allowed for this specific cart line, accounting for other variant lines of the same product already in the cart */
  maxQuantity: number;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  variantId?: number | null;
  /** @nullable */
  variantLabel?: string | null;
}